fix: select month end error for scheduled backup
